Added new hive client subclass, Removed config in Cachingmetastore
