+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveClientConfig.java
@@ -351,76 +351,6 @@ public class HiveClientConfig
         return this;
     }
 
-    /** twitter specified config */
-    @Min(0)
-    public int getMaxMetastoreRetryAttempts()
-    {
-        return maxMetastoreRetryAttempts;
-    }
-
-    @Config("hive.metastore.client.create.max-retry-attempts")
-    public HiveClientConfig setMaxMetastoreRetryAttempts(int maxMetastoreRetryAttempts)
-    {
-        this.maxMetastoreRetryAttempts = maxMetastoreRetryAttempts;
-        return this;
-    }
-
-    @Min(1)
-    public double getMetastoreRetryScaleFactor()
-    {
-        return metastoreRetryScaleFactor;
-    }
-
-    @Config("hive.metastore.client.create.retry-scale-fator")
-    public HiveClientConfig setMetastoreRetryScaleFactor(double metastoreRetryScaleFactor)
-    {
-        this.metastoreRetryScaleFactor = metastoreRetryScaleFactor;
-        return this;
-    }
-
-    @NotNull
-    @MinDuration("1ms")
-    public Duration getMaxMetastoreRetryTime()
-    {
-        return maxMetastoreRetryTime;
-    }
-
-    @Config("hive.metastore.client.create.max-retry-time")
-    public HiveClientConfig setMaxMetastoreRetryTime(Duration maxMetastoreRetryTime)
-    {
-        this.maxMetastoreRetryTime = maxMetastoreRetryTime;
-        return this;
-    }
-
-    @NotNull
-    @MinDuration("1ms")
-    public Duration getMinMetastoreRetrySleepTime()
-    {
-        return minMetastoreRetrySleepTime;
-    }
-
-    @Config("hive.metastore.client.create.min-retry-sleep-time")
-    public HiveClientConfig setMinMetastoreRetrySleepTime(Duration minMetastoreRetrySleepTime)
-    {
-        this.minMetastoreRetrySleepTime = minMetastoreRetrySleepTime;
-        return this;
-    }
-
-    @NotNull
-    @MinDuration("1ms")
-    public Duration getMaxMetastoreRetrySleepTime()
-    {
-        return maxMetastoreRetrySleepTime;
-    }
-
-    @Config("hive.metastore.client.create.max-retry-sleep-time")
-    public HiveClientConfig setMaxMetastoreRetrySleepTime(Duration maxMetastoreRetrySleepTime)
-    {
-        this.maxMetastoreRetrySleepTime = maxMetastoreRetrySleepTime;
-        return this;
-    }
-    /** end twitter */
-
     @NotNull
     public Duration getMetastoreCacheTtl()
     {
