+++ presto-hive/src/main/java/com/facebook/presto/twitter/hive/HiveClientModuleRetryConfigurable.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.twitter.hive;
+
+import com.facebook.presto.hive.HiveClientModule
+import com.facebook.presto.hive.metastore.HiveMetastore;
+import com.facebook.presto.spi.NodeManager;
+import com.facebook.presto.spi.PageIndexerFactory;
+import com.facebook.presto.spi.type.TypeManager;
+import com.google.inject.Binder;
+
+import static io.airlift.configuration.ConfigBinder.configBinder;
+
+public class HiveClientModuleRetryConfigurable
+        extends HiveClientModule
+{
+    public HiveClientModuleRetryConfigurable(
+        String connectorId, HiveMetastore metastore, TypeManager typeManager, PageIndexerFactory pageIndexerFactory, NodeManager nodeManager)
+    {
+        super(connectorId, metastore, typeManager, pageIndexerFactory, nodeManager)
+    }
+
+    @Override
+    public void configure(Binder binder)
+    {
+        configBinder(binder).bindConfig(HiveClientRetryConfig.class);
+        super.configure(binder)
+    }
+}
